# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: RiskMind API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List signals
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, triaged, finding, dismissed]
        - name: source
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Paginated list of signals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalListResponse"
    post:
      operationId: createSignal
      tags: [signals]
      summary: Create signal (single or bulk)
      description: Send a single signal object or an array of up to 100 signals
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/CreateSignalRequest"
                - type: array
                  items:
                    $ref: "#/components/schemas/CreateSignalRequest"
                  maxItems: 100
      responses:
        "201":
          description: Signal(s) created
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Signal"
                  - $ref: "#/components/schemas/BulkSignalResponse"
  /v1/signals/{id}:
    get:
      operationId: getSignal
      tags: [signals]
      summary: Get signal by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Signal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/signals/{id}/status:
    patch:
      operationId: updateSignalStatus
      tags: [signals]
      summary: Transition signal status
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSignalStatusRequest"
      responses:
        "200":
          description: Signal status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/signals/{signalId}/promote:
    post:
      operationId: promoteSignalToFinding
      tags: [signals, findings]
      summary: Promote signal to finding
      security:
        - bearerAuth: []
      parameters:
        - name: signalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoteSignalRequest"
      responses:
        "201":
          description: Finding created from signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/findings:
    get:
      operationId: listFindings
      tags: [findings]
      summary: List findings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, investigating, resolved, false_positive]
        - name: riskId
          in: query
          schema:
            type: string
            format: uuid
        - name: vendorId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Paginated list of findings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingListResponse"
    post:
      operationId: createFinding
      tags: [findings]
      summary: Create finding
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFindingRequest"
      responses:
        "201":
          description: Finding created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
  /v1/findings/{id}:
    get:
      operationId: getFinding
      tags: [findings]
      summary: Get finding by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Finding details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateFinding
      tags: [findings]
      summary: Update finding
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFindingRequest"
      responses:
        "200":
          description: Finding updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
...
  /v1/risks/{riskId}/enrich:
    post:
      operationId: enrichRiskDescription
      tags: [aiEnrichment]
      summary: Trigger AI enrichment of risk description
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/riskId"
      responses:
        "202":
          description: Enrichment job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
  /v1/vendors/{vendorId}/documents/{documentId}/summarize:
    post:
      operationId: summarizeVendorDocument
      tags: [aiEnrichment]
      summary: Trigger AI summarization of vendor document
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/vendorId"
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "202":
          description: Summarization job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
  /v1/jobs/{id}:
    get:
      operationId: getJobStatus
      tags: [jobs]
      summary: Get job status
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    riskId:
      name: riskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    vendorId:
      name: vendorId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RFC7807Error"
    Conflict:
      description: Conflict - invalid state transition or duplicate
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RFC7807Error"
  schemas:
...
    RFC7807Error:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
      required:
        - type
        - title
        - status
...
    SignalStatus:
      type: string
      enum: [pending, triaged, finding, dismissed]
    Signal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        source:
          type: string
        content:
          type: string
        status:
          $ref: "#/components/schemas/SignalStatus"
        classification:
          type: string
          nullable: true
        confidence:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSignalRequest:
      type: object
      required: [source, content]
      properties:
        source:
          type: string
        content:
          type: string
    UpdateSignalStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/SignalStatus"
        classification:
          type: string
        confidence:
          type: number
    BulkSignalResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
        count:
          type: integer
    SignalListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PromoteSignalRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        riskId:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
    FindingStatus:
      type: string
      enum: [open, investigating, resolved, false_positive]
    Finding:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        signalId:
          type: string
          format: uuid
          nullable: true
        riskId:
          type: string
          format: uuid
          nullable: true
        vendorId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/FindingStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateFindingRequest:
      type: object
      required: [title]
      properties:
        signalId:
          type: string
          format: uuid
        riskId:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
    UpdateFindingRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/FindingStatus"
        riskId:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
    FindingListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Finding"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    JobStatus:
      type: string
      enum: [pending, processing, completed, failed, dead]
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
          nullable: true
        queue:
          type: string
        type:
          type: string
        payload:
          type: object
        result:
          type: object
          nullable: true
        status:
          $ref: "#/components/schemas/JobStatus"
        attempts:
          type: integer
        maxAttempts:
          type: integer
        lastError:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    JobAccepted:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
        message:
          type: string