# lib/api-zod/src/generated/types/jobAccepted.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RiskMind API specification
 * OpenAPI spec version: 0.3.0
 */

export interface JobAccepted {
  jobId?: string;
  status?: string;
  message?: string;
}